设计manager统一管理

+++ performance/src/main/java/com/xbl/performance/Config.java
@@ -1,8 +1,8 @@
 package com.xbl.performance;
 
 public class Config {
-    public static int BLOCK_WATCHER_INTERVAL = 1 * 1000;
-    public static int FPS_WATCHER_INTERVAL = 1 * 1000;
+    public static long BLOCK_WATCHER_INTERVAL = 1 * 1000;
+    public static long FPS_WATCHER_INTERVAL = 1 * 1000;
 
     public static final String TYPE_BLOCK = "type_block";
